agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = ""
authors = [
    {name = "orca-copilotkit",email = "testapp@copilotkit.ai"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "copilotkit (>=0.1.52,<0.2.0)",
    "ag-ui-protocol (>=0.1.7,<0.2.0)",
    "fastapi (>=0.115.14,<0.116.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "yfinance (>=0.2.64,<0.3.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "lxml (>=6.0.0,<7.0.0)",
    "llama-index-core (>=0.12.50,<0.13.0)",
    "llama-index-agent-openai (>=0.4.12,<0.5.0)",
    "openai (>=1.97.0,<2.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
